+++ src/main/java/com/invadermonky/magicultureintegrations/core/mixins/industrialcraft/BlockTileEntityHarvestableMixin.java
@@ -1,0 +1,45 @@
+package com.invadermonky.magicultureintegrations.core.mixins.industrialcraft;
+
+import com.invadermonky.magicultureintegrations.api.block.IHarvestableCrop;
+import ic2.api.crops.ICropTile;
+import ic2.core.block.BlockTileEntity;
+import ic2.core.crop.TileEntityCrop;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.item.ItemStack;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.NonNullList;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+import org.spongepowered.asm.mixin.Mixin;
+
+@Mixin(value = BlockTileEntity.class, remap = false)
+public class BlockTileEntityHarvestableMixin implements IHarvestableCrop {
+
+    @Override
+    public BlockPos getHarvestPosition(World world, BlockPos cropPos) {
+        return cropPos;
+    }
+
+    @Override
+    public @NotNull HarvestResult getHarvestResult(World world, BlockPos pos) {
+        TileEntity tile = world.getTileEntity(pos);
+        if (tile instanceof TileEntityCrop) {
+            TileEntityCrop crop = (TileEntityCrop) tile;
+            return crop.getCrop() != null && crop.getCrop().canBeHarvested(crop) ? HarvestResult.HARVEST : HarvestResult.CLAIM;
+        }
+        return HarvestResult.PASS;
+    }
+
+    @Override
+    public @NotNull NonNullList<ItemStack> harvestCrop(@Nullable EntityPlayer player, World world, BlockPos pos, boolean silkTouch, int fortune) {
+        NonNullList<ItemStack> drops = NonNullList.create();
+        TileEntity tile = world.getTileEntity(pos);
+        if (tile instanceof ICropTile && getHarvestResult(world, pos) == HarvestResult.HARVEST) {
+            drops.addAll(((ICropTile) tile).performHarvest());
+        }
+        return drops;
+    }
+
+}